+++ src/main/java/com/polymarket/model/RfqUserQuote.java
@@ -3,6 +3,8 @@ package com.polymarket.model;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 /** Input for creating an RFQ quote. Mirrors TS {@code RfqUserQuote}. */
 @Value
 @Builder
@@ -10,7 +12,7 @@ public class RfqUserQuote {
 
     String requestId;
     String tokenID;
-    double price;
-    double size;
+    BigDecimal price;
+    BigDecimal size;
     Side side;
 }
